Assert Cursor#close() for Get<List> Operation, improvements for Get Operation tests
