+++ storio-sqlite/src/test/java/com/pushtorefresh/storio/sqlite/operation/get/PreparedGetTest.java
@@ -2,24 +2,17 @@ package com.pushtorefresh.storio.sqlite.operation.get;
 
 import android.database.Cursor;
 
-import com.pushtorefresh.storio.sqlite.StorIOSQLite;
-import com.pushtorefresh.storio.sqlite.query.Query;
-import com.pushtorefresh.storio.sqlite.query.RawQuery;
-
 import org.junit.Test;
-import org.mockito.invocation.InvocationOnMock;
-import org.mockito.stubbing.Answer;
 
-import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicLong;
 
-import static junit.framework.Assert.assertEquals;
-import static junit.framework.Assert.assertSame;
-import static org.mockito.Mockito.mock;
+import rx.Observable;
+
+import static org.junit.Assert.assertNotNull;
+import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
 public class PreparedGetTest {
 
@@ -57,34 +50,67 @@ public class PreparedGetTest {
     public void getCursorObservable() {
         final GetStub getStub = new GetStub();
 
-        final Cursor cursor = getStub.storIOSQLite
+        final Observable<Cursor> cursorObservable = getStub.storIOSQLite
                 .get()
                 .cursor()
                 .withQuery(getStub.query)
                 .withGetResolver(getStub.getResolverForCursor)
                 .prepare()
-                .createObservable()
-                .toBlocking()
-                .last();
+                .createObservable();
 
-        getStub.verifyQueryBehaviorForCursor(cursor);
+        getStub.verifyQueryBehaviorForCursor(cursorObservable);
     }
+
+    @Test
+    public void getCursorObservableStreamTakeOne() {
+        final GetStub getStub = new GetStub();
+
+        final Observable<Cursor> cursorObservable = getStub.storIOSQLite
+                .get()
+                .cursor()
+                .withQuery(getStub.query)
+                .withGetResolver(getStub.getResolverForCursor)
+                .prepare()
+                .createObservableStream()
+                .take(1);
+
+        getStub.verifyQueryBehaviorForCursor(cursorObservable);
+
+        verify(getStub.storIOSQLite, times(1)).observeChangesInTables(eq(Collections.singleton(getStub.query.table)));
+    }
+
 
     @Test
     public void getListOfObjectsObservable() {
         final GetStub getStub = new GetStub();
 
-        final List<TestItem> testItems = getStub.storIOSQLite
+        final Observable<List<TestItem>> testItemsObservable = getStub.storIOSQLite
                 .get()
                 .listOfObjects(TestItem.class)
                 .withQuery(getStub.query)
                 .withGetResolver(getStub.getResolverForObject)
                 .prepare()
-                .createObservable()
-                .toBlocking()
-                .last();
+                .createObservable();
 
-        getStub.verifyQueryBehaviorForList(testItems);
+        getStub.verifyQueryBehaviorForList(testItemsObservable);
+    }
+
+    @Test
+    public void getListOfObjectsObservableStreamTakeOne() {
+        final GetStub getStub = new GetStub();
+
+        final Observable<List<TestItem>> testItemsObservable = getStub.storIOSQLite
+                .get()
+                .listOfObjects(TestItem.class)
+                .withQuery(getStub.query)
+                .withGetResolver(getStub.getResolverForObject)
+                .prepare()
+                .createObservableStream()
+                .take(1);
+
+        getStub.verifyQueryBehaviorForList(testItemsObservable);
+
+        verify(getStub.storIOSQLite, times(1)).observeChangesInTables(eq(Collections.singleton(getStub.query.table)));
     }
 
     @Test
@@ -106,17 +132,35 @@ public class PreparedGetTest {
     public void getCursorWithRawQueryObservable() {
         final GetStub getStub = new GetStub();
 
-        final Cursor cursor = getStub.storIOSQLite
+        final Observable<Cursor> cursorObservable = getStub.storIOSQLite
                 .get()
                 .cursor()
                 .withQuery(getStub.rawQuery)
                 .withGetResolver(getStub.getResolverForCursor)
                 .prepare()
-                .createObservable()
-                .toBlocking()
-                .last();
+                .createObservable();
 
-        getStub.verifyRawQueryBehaviorForCursor(cursor);
+        getStub.verifyRawQueryBehaviorForCursor(cursorObservable);
+    }
+
+    @Test
+    public void getCursorWithRawQueryObservableStreamTakeOne() {
+        final GetStub getStub = new GetStub();
+
+        final Observable<Cursor> cursorObservable = getStub.storIOSQLite
+                .get()
+                .cursor()
+                .withQuery(getStub.rawQuery)
+                .withGetResolver(getStub.getResolverForCursor)
+                .prepare()
+                .createObservableStream()
+                .take(1);
+
+        getStub.verifyRawQueryBehaviorForCursor(cursorObservable);
+
+        assertNotNull(getStub.rawQuery.affectedTables);
+
+        verify(getStub.storIOSQLite, times(1)).observeChangesInTables(getStub.rawQuery.affectedTables);
     }
 
     @Test
@@ -138,141 +182,35 @@ public class PreparedGetTest {
     public void getListOfObjectsWithRawQueryObservable() {
         final GetStub getStub = new GetStub();
 
-        final List<TestItem> testItems = getStub.storIOSQLite
+        final Observable<List<TestItem>> testItemsObservable = getStub.storIOSQLite
                 .get()
                 .listOfObjects(TestItem.class)
                 .withQuery(getStub.rawQuery)
                 .withGetResolver(getStub.getResolverForObject)
                 .prepare()
-                .createObservable()
-                .toBlocking()
-                .last();
+                .createObservable();
 
-        getStub.verifyRawQueryBehaviorForList(testItems);
+        getStub.verifyRawQueryBehaviorForList(testItemsObservable);
     }
 
-    private static class TestItem {
+    @Test
+    public void getListOfObjectsWithRawQueryObservableStreamTakeOne() {
+        final GetStub getStub = new GetStub();
 
-        private static final AtomicLong COUNTER = new AtomicLong(0);
+        final Observable<List<TestItem>> testItemsObservable = getStub.storIOSQLite
+                .get()
+                .listOfObjects(TestItem.class)
+                .withQuery(getStub.rawQuery)
+                .withGetResolver(getStub.getResolverForObject)
+                .prepare()
+                .createObservableStream()
+                .take(1);
 
-        private Long id = COUNTER.incrementAndGet();
+        getStub.verifyRawQueryBehaviorForList(testItemsObservable);
 
-        public Long getId() {
-            return id;
-        }
+        assertNotNull(getStub.rawQuery.affectedTables);
 
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-
-            TestItem testItem = (TestItem) o;
-
-            return !(id != null ? !id.equals(testItem.id) : testItem.id != null);
-        }
-
-        @Override
-        public int hashCode() {
-            return id != null ? id.hashCode() : 0;
-        }
+        verify(getStub.storIOSQLite, times(1)).observeChangesInTables(getStub.rawQuery.affectedTables);
     }
 
-    private static class GetStub {
-        final StorIOSQLite storIOSQLite;
-        final Query query;
-        final RawQuery rawQuery;
-        final GetResolver<TestItem> getResolverForObject;
-        final GetResolver<Cursor> getResolverForCursor;
-        final Cursor cursor;
-        final List<TestItem> testItems;
-        private final StorIOSQLite.Internal internal;
-
-        @SuppressWarnings("unchecked")
-        GetStub() {
-            storIOSQLite = mock(StorIOSQLite.class);
-            internal = mock(StorIOSQLite.Internal.class);
-
-            when(storIOSQLite.internal())
-                    .thenReturn(internal);
-
-            query = new Query.Builder()
-                    .table("test_table")
-                    .build();
-
-            rawQuery = new RawQuery.Builder()
-                    .query("test sql")
-                    .build();
-
-            getResolverForObject = mock(GetResolver.class);
-            getResolverForCursor = mock(GetResolver.class);
-            cursor = mock(Cursor.class);
-
-            testItems = new ArrayList<TestItem>();
-            testItems.add(new TestItem());
-            testItems.add(new TestItem());
-            testItems.add(new TestItem());
-
-            when(cursor.moveToNext()).thenAnswer(new Answer<Boolean>() {
-                int invocationsCount = 0;
-
-                @Override
-                public Boolean answer(InvocationOnMock invocation) throws Throwable {
-                    return invocationsCount++ < testItems.size();
-                }
-            });
-
-            when(storIOSQLite.get())
-                    .thenReturn(new PreparedGet.Builder(storIOSQLite));
-
-            when(getResolverForObject.performGet(storIOSQLite, query))
-                    .thenReturn(cursor);
-
-            when(getResolverForObject.performGet(storIOSQLite, rawQuery))
-                    .thenReturn(cursor);
-
-            when(getResolverForCursor.performGet(storIOSQLite, query))
-                    .thenReturn(cursor);
-
-            when(getResolverForCursor.performGet(storIOSQLite, rawQuery))
-                    .thenReturn(cursor);
-
-            when(getResolverForObject.mapFromCursor(cursor))
-                    .thenAnswer(new Answer<TestItem>() {
-                        int invocationsCount = 0;
-
-                        @Override
-                        public TestItem answer(InvocationOnMock invocation) throws Throwable {
-                            final TestItem testItem = testItems.get(invocationsCount);
-                            invocationsCount++;
-                            return testItem;
-                        }
-                    });
-        }
-
-        private void verifyQueryBehaviorForCursor(Cursor actualCursor) {
-            verify(storIOSQLite, times(1)).get();
-            verify(getResolverForCursor, times(1)).performGet(storIOSQLite, query);
-            assertSame(cursor, actualCursor);
-        }
-
-        private void verifyQueryBehaviorForList(List<TestItem> actualList) {
-            verify(storIOSQLite, times(1)).get();
-            verify(getResolverForObject, times(1)).performGet(storIOSQLite, query);
-            verify(getResolverForObject, times(testItems.size())).mapFromCursor(cursor);
-            assertEquals(testItems, actualList);
-        }
-
-        private void verifyRawQueryBehaviorForCursor(Cursor actualCursor) {
-            verify(storIOSQLite, times(1)).get();
-            verify(getResolverForCursor, times(1)).performGet(storIOSQLite, rawQuery);
-            assertSame(cursor, actualCursor);
-        }
-
-        private void verifyRawQueryBehaviorForList(List<TestItem> actualList) {
-            verify(storIOSQLite, times(1)).get();
-            verify(getResolverForObject, times(1)).performGet(storIOSQLite, rawQuery);
-            verify(getResolverForObject, times(testItems.size())).mapFromCursor(cursor);
-            assertEquals(testItems, actualList);
-        }
-    }
 }
